src/app/home/home.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';

import { HomeComponent } from './home.component';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import {HttpClientModule} from '@angular/common/http';
import { RegisterService } from '../register.service';

describe('HomeComponent', () => {
  let component: HomeComponent;
  let fixture: ComponentFixture<HomeComponent>;

  // beforeEach(async () => {
  //   await TestBed.configureTestingModule({
  //     declarations: [ HomeComponent ]
  //   })
  //   .compileComponents();
  // });

  // beforeEach(() => {
  //   fixture = TestBed.createComponent(HomeComponent);
  //   component = fixture.componentInstance;
  //   fixture.detectChanges();
  // });

  // it('should create', () => {
  //   expect(component).toBeTruthy();
  // });

  describe('RegisterService', () => {

    beforeEach(() => TestBed.configureTestingModule({
      imports: [HttpClientTestingModule], 
      providers: [RegisterService]
    }));

     it('should be created', () => {
      const service: RegisterService = TestBed.get(RegisterService);
      expect(service).toBeTruthy();
     });

     it('should have getData function', () => {
      const service: RegisterService = TestBed.get(RegisterService);
      expect(service.getData).toBeTruthy();
     });

  });
});
